Use the right format for printing out total bytes.  (Having format checking with notice is a good thing.)

+++ backends/batcher.c
@@ -101,8 +101,8 @@ RequeueAndExit(off_t Cookie, char *line, long BytesInArt)
 
     notice("batcher %s times user %.3f system %.3f elapsed %.3f",
            Host, usertime, systime, STATend - STATbegin);
-    notice("batcher %s stats batches %d articles %d bytes %ld",
-           Host, BATCHcount, ArtsWritten, BytesWritten);
+    notice("batcher %s stats batches %d articles %d bytes %lu",
+           Host, BATCHcount, ArtsWritten, (unsigned long) BytesWritten);
 
     /* Last batch exit okay? */
     if (BATCHstatus == 0) {
